requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==5.0.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.0.4
colorama==0.4.4
contourpy==1.0.5
cycler==0.12.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.44.0
idna==3.4
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyter-server==1.23.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.0
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.3
mistune==2.0.4
mpmath==1.3.0
msgpack==1.0.6
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nest-asyncio==1.5.4
networkx==3.1
notebook_shim==0.2.2
notebook==6.5.2
numba==0.57.1
numpy==1.24.4
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
platformdirs==2.5.1
ply==3.11
pooch==1.8.0
prometheus-client==0.15.0
prompt-toolkit==3.0.28
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==303
pywinpty==2.0.9
pyzmq==22.3.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.0
setuptools==68.0.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
stack-data==0.2.0
sympy==1.11.1
terminado==0.17.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
tornado==6.1
traitlets==5.5.0
typing_extensions==4.7.1
unicodedata2==15.1.0
urllib3==2.0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
wheel==0.38.4
win-inet-pton==1.1.0
